+++ src/main/java/nextstep/web/ReservationQueryingDAO.java
@@ -1,7 +1,7 @@
 package nextstep.web;
 
-import nextstep.Reservation;
-import nextstep.Theme;
+import nextstep.domain.Reservation;
+import nextstep.domain.Theme;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.RowMapper;
 import org.springframework.stereotype.Repository;
