+++ src/main/java/nextstep/web/ReservationUpdatingDAO.java
@@ -1,6 +1,6 @@
 package nextstep.web;
 
-import nextstep.Reservation;
+import nextstep.domain.Reservation;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.support.GeneratedKeyHolder;
 import org.springframework.jdbc.support.KeyHolder;
