+++ src/test/java/nextstep/web/ReservationQueryingDAOTest.java
@@ -1,6 +1,6 @@
 package nextstep.web;
 
-import nextstep.Reservation;
+import nextstep.domain.Reservation;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
