+++ src/test/java/nextstep/web/ReservationUpdatingDAOTest.java
@@ -1,7 +1,7 @@
 package nextstep.web;
 
-import nextstep.Reservation;
-import nextstep.Theme;
+import nextstep.domain.Reservation;
+import nextstep.domain.Theme;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
